Move component Couter to its folder. Update imports. Add tests for Counter
